// frontend/src/Types.ts
export type Current_station = {
  codec: string
  bitrate: number
  station_id: string
  favorite: boolean
  language: string
  favicon: string
  tags: string
  state: string
  home_page: string
  country_codes: string
  country: string
  loading: boolean
  name: string
  paused: boolean
  url_resolved: string
  votes: number
}